Create middleware and routers
