+++ middleware/logger.js
@@ -1,0 +1,6 @@
+export default (req, _, next) => {
+    if (process.env.NODE_ENV === 'dev') {
+        console.log(req.query);
+    }
+    next();
+};
