+++ routers/index.js
@@ -1,0 +1,7 @@
+import router from './router';
+import testRouter from './testRouter';
+
+export {
+    router,
+    testRouter
+};
