+++ server.js
@@ -1,12 +1,11 @@
 import express from 'express';
 import mongoose from 'mongoose';
-import { getWordData, getWords } from './controllers/words';
-import { seedDatabase } from './dictionaries/seed';
+import { testRouter } from './routers';
+import logger from './middleware/logger';
 import { SERVER_PORT, MONGO_URI, TEST_MONGO_URI } from './config';
 
 const app = express();
 const router = express.Router();
-const testRouter = express.Router();
 
 mongoose.connect(process.env.NODE_ENV === 'test' ? TEST_MONGO_URI : MONGO_URI, {
     useNewUrlParser: true,
@@ -23,21 +22,7 @@ app.get('/', (_, res) => {
     res.send('Hello World!');
 });
 
-router.get('/', (_, res) => {
-    res.send('Welcome to the Igbo English Dictionary API');
-});
-
-router.get('/words', getWordData);
-
-testRouter.post('/populate', seedDatabase);
-testRouter.get('/words', getWords);
-
-app.use('*', (req, _, next) => {
-    if (process.env.NODE_ENV === 'dev') {
-        console.log(req.query);
-    }
-    next();
-});
+app.use('*', logger);
 
 /* Grabs data from JSON dictionary */
 app.use('/api/v1/search', router);
